requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.6
django-crispy-forms==2.1
django-phonenumber-field==7.2.0
install==1.3.5
phonenumberslite==8.13.24
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
